basic specifications of cards are set

+++ cpp files/Object.cpp
@@ -4,7 +4,7 @@
 
 #include <header files/Object.h>
 
-Object::Object(int tp, int trgt, double hs, int hp, int d, double r, int tr, QTimer *timer)
+Object::Object(int tp, int trgt, double hs, int hp, int d, double r, double tr, QTimer *timer)
 {
     type = tp;
     target = trgt;
